refactor: move add screen from menu to screens
