+++ src/components/SegmentMenu/SegmentMenu.tsx
@@ -6,7 +6,6 @@ import {
   CopyAll,
   Delete,
   MoreVert,
-  QueuePlayNext,
 } from '@mui/icons-material'
 import {
   Divider,
@@ -20,7 +19,6 @@ import {
 
 import { useProgram } from '../../hooks'
 import { SegmentMediaType } from '../../types'
-import { generateNewId } from '../../utils'
 import { Popup } from '../Popup'
 
 export type SegmentMenuProps = {
@@ -28,10 +26,9 @@ export type SegmentMenuProps = {
   media?: SegmentMediaType
 }
 export const SegmentMenuRaw = ({ segmentId, media }: SegmentMenuProps) => {
-  const { removeSegment, addNextSegmentById, addScreenToMedia } = useProgram()
+  const { removeSegment, addNextSegmentById } = useProgram()
 
   const [showPasteId, setShowPasteId] = useState<boolean>(false)
-  const [showAddScreen, setShowAddScreen] = useState<boolean>(false)
   const [showDeleteSegment, setShowDeleteSegment] = useState<boolean>(false)
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
@@ -53,28 +50,6 @@ export const SegmentMenuRaw = ({ segmentId, media }: SegmentMenuProps) => {
     setShowPasteId(false)
     handleClose()
   }, [addNextSegmentById, segmentId])
-
-  const screenTitleRef = useRef<HTMLInputElement>()
-
-  const handleAddScreen = useCallback(
-    (e: any) => {
-      if (!media) return
-      const screen = {
-        id: generateNewId(),
-        title: screenTitleRef.current?.value || 'screen title',
-        mediaSrc: `/programMedia/${e.target.files[0].name}`,
-      }
-      addScreenToMedia(media.id, screen)
-      setShowAddScreen(false)
-      handleClose()
-    },
-    [addScreenToMedia, media]
-  )
-
-  const handleShowAddScreen = useCallback(() => {
-    setShowAddScreen(true)
-    handleClose()
-  }, [])
 
   const handlePasteId = useCallback(() => {
     setShowPasteId(true)
@@ -118,14 +93,6 @@ export const SegmentMenuRaw = ({ segmentId, media }: SegmentMenuProps) => {
           'aria-labelledby': 'basic-button',
         }}
       >
-        {media && (
-          <MenuItem>
-            <ListItemIcon onClick={handleShowAddScreen}>
-              <QueuePlayNext fontSize="small" />
-            </ListItemIcon>
-            <ListItemText>Add Screen</ListItemText>
-          </MenuItem>
-        )}
         <MenuItem onClick={handlePasteId}>
           <ListItemIcon>
             <ContentPaste fontSize="small" />
@@ -146,28 +113,7 @@ export const SegmentMenuRaw = ({ segmentId, media }: SegmentMenuProps) => {
           <ListItemText>Delete Segment</ListItemText>
         </MenuItem>
       </Menu>
-      {media && showAddScreen && (
-        <Popup
-          onClose={() => setShowAddScreen(false)}
-          header="Add screen"
-          bodyText="First give it a title, and then select the video file:"
-        >
-          <TextField
-            inputRef={screenTitleRef}
-            placeholder={'screen title'}
-            size="small"
-          />
-          <IconButton component="label">
-            <input
-              hidden
-              accept="video/mp4"
-              type="file"
-              onChange={handleAddScreen}
-            />
-            <QueuePlayNext />
-          </IconButton>
-        </Popup>
-      )}
+
       {showDeleteSegment && (
         <Popup
           onClose={() => setShowDeleteSegment(false)}
